+++ internal/manifest/procfile/generate.go
@@ -3,6 +3,7 @@ package procfile
 import (
 	"fmt"
 	"io"
+	"sort"
 
 	"gopkg.in/alessio/shellescape.v1"
 	"mvdan.cc/sh/v3/syntax"
@@ -21,7 +22,16 @@ func Generate(w io.Writer, procs []Process) error {
 			cmd.Args = append(cmd.Args, literalWord(arg))
 		}
 
-		for key, val := range proc.Environment {
+		keys := make([]string, len(proc.Environment))
+		i := 0
+		for key := range proc.Environment {
+			keys[i] = key
+			i++
+		}
+		sort.Strings(keys)
+
+		for _, key := range keys {
+			val := proc.Environment[key]
 			cmd.Assigns = append(cmd.Assigns, &syntax.Assign{
 				Name:  &syntax.Lit{Value: key},
 				Value: literalWord(val),